Moved orthodb utils from pipeline.py to orthodb.py, removed orthodb cli parsers, added statistics of methods.
